Extract C# object serialization methods to a dedicated class, allowing AssertException to become more focused and to better reveal which few such methods are actual public entry points to the broader serialization scheme.
